condenses chatgpt script to one file so we are getting backstory and name there, now renders both on output page. input pseudocode to prep file for using db query to fetch parameters and create prompts
