+++ views/layouts/main.handlebars
@@ -11,7 +11,7 @@
 
   <body>
     <header>
-      <h1>Welcome to LegendSmith!</h1>
+      <h1 id="pagetitle">Welcome to LegendSmith!</h1>
     </header>
         <nav>
           <a href="/">Home</a>
