+++ views/partials/output.handlebars
@@ -1,8 +1,9 @@
 {{#if loggedIn}}
   <div>
+    <p id='charnameappend'>Character Name:</p>
     <p id='backstoryappend'>Backstory:</p>
     {{! maybe add ... (for the user to continue the story) }}
   </div>
 {{/if}}
 
-<script src='/js/chatgptscriptgetbackstory.js'></script>
+<script src='/js/chatgptscript.js'></script>
